snapshots/snapshot_scheduler/src/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.7
# by the following command:
#
#    pip-compile --resolver=backtracking ./requirements.in
#
anyio==3.6.2
    # via httpcore
attrs==20.3.0
    # via -r ./requirements.in
boto3==1.7.31
    # via wellcome-aws-utils
botocore==1.10.31
    # via
    #   boto3
    #   s3transfer
certifi==2022.12.7
    # via
    #   httpcore
    #   httpx
daiquiri==1.3.0
    # via wellcome-aws-utils
docutils==0.14
    # via botocore
h11==0.14.0
    # via httpcore
httpcore==0.16.3
    # via httpx
httpx==0.23.3
    # via -r ./requirements.in
idna==3.4
    # via
    #   anyio
    #   rfc3986
jmespath==0.9.3
    # via
    #   boto3
    #   botocore
python-dateutil==2.7.3
    # via
    #   botocore
    #   wellcome-aws-utils
rfc3986[idna2008]==1.5.0
    # via httpx
s3transfer==0.1.13
    # via boto3
six==1.11.0
    # via python-dateutil
sniffio==1.3.0
    # via
    #   anyio
    #   httpcore
    #   httpx
typing-extensions==4.4.0
    # via
    #   anyio
    #   h11
wellcome-aws-utils==2.1.1
    # via -r ./requirements.in
